requirements.txt
arrow==0.12.1
asn1crypto==0.24.0
beautifulsoup4==4.6.3
bibtexparser==1.0.1
boto3==1.7.80
botocore==1.10.80
certifi==2018.8.13
cffi==1.11.5
chardet==3.0.4
crossrefapi==1.3.0
cryptography==2.3.1
Django==1.11.1
django-background-tasks==1.1.13
django-compat==1.0.15
django-selenium==0.9.8
docutils==0.14
future==0.16.0
idna==2.7
jmespath==0.9.3
lxml==4.2.4
pycparser==2.18
pyOpenSSL==18.0.0
pyparsing==2.2.0
PyPDF2==1.26.0
PySocks==1.6.8
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
retrying==1.3.3
s3transfer==0.1.13
scholarly==0.2.3
selenium==3.14.0
six==1.11.0
urllib3==1.23
